requirements.txt
fastapi
httpx
redis
uvicorn
numpy==1.26.0
aioredis
aiohttp
openai
pydantic
beautifulsoup4
requests
python-jose
pandas
langchain
faiss-cpu
sentence_transformers
psutil
passlib
python-dotenv
bcrypt
locust
rank-bm25
nltk
gunicorn
python-box==6.0.2
ensure
pytest
pytest-asyncio
xgboost
evaluate
torch==2.2.2
transformers==4.39.3
bert_score==0.3.13
aiofiles
imblearn
rouge-score
tabulate
seaborn
matplotlib
streamlit
nest_asyncio